requirements.txt
python-dotenv==0.10.3
python-telegram-bot==12.3.0
redis==3.3.11
vk_api==11.8.0